Define priority for video threads
Video threads have soft realtime requirements that
must be met in order to maintain reliable frame delivery
even when the system is under high load. This CL
defines a new priority that can be used by video
threads, enabling them to be scheduled appropriately
relative to other system threads.

Change-Id: Idd57207e30309dfdff24389db0acf107532f9e5a
related-to-bug:63898080